Doing collectd query in separate process.
There are a few issues with collectd and locking the hypervisor database.
It's not clear what code path this is occuring on (or if it's even a vms
problem or a collectd problem) but this should side-step the issue by
safely forking a separate python process in order to do the query.